#355 Update FloatingTexts

+++ mdc/DataCenter/MDC/Data/Dtos/PositionDto.cs
@@ -6,10 +6,10 @@ namespace MDC.Data.Dtos
     {
         public string Level { get; set; }
 
-        public double X { get; set; }
+        public int X { get; set; }
 
-        public double Y { get; set; }
+        public int Y { get; set; }
 
-        public double Z { get; set; }
+        public int Z { get; set; }
     }
 }
